+++ test/run_tests.m
@@ -1,8 +1,11 @@
 verify_test_data();
 % Updates the required euphonic versions
 curdir = split(fileparts(mfilename('fullpath')), filesep);
-append(py.sys.path, char(join(curdir(1:end-1), filesep)));
+repodir = char(join(curdir(1:end-1), filesep));
+disp(['Adding ', repodir, ' to Python path']);
+append(py.sys.path, repodir);
 py.euphonic_version.update_euphonic_version();
+
 res = runtests('test/EuphonicTest.m', 'Tag', 'integration');
 passed = [res.Passed];
 if ~all(passed)